@@ -1,25 +1,24 @@
 {
-    "name": "torrent-organizer",
-    "version": "1.0.0",
-    "private": true,
-    "main": "parse.js",
-    "scripts": {
-        "start": "webpack --watch --progress"
-    },
-    "author": "\"Lord-Lavios\"",
-    "license": "MIT",
-    "dependencies": {
-        "babel-eslint": "^7.1.1",
-        "babel-loader": "^6.3.2",
-        "babel-plugin-transform-async-to-generator": "^6.22.0",
-        "babel-plugin-transform-runtime": "^6.22.0",
-        "babel-polyfill": "^6.22.0",
-        "fs": "^0.0.1-security",
-        "webpack": "^2.2.1"
-    },
-    "devDependencies": {
-        "babel-preset-es2015": "^6.22.0",
-        "babel-preset-stage-3": "^6.22.0",
-        "eslint": "^3.15.0"
-    }
+	"name": "Torrent-Organizer",
+	"version": "1.0.0",
+	"main": "src/index.js",
+	"scripts": {
+		"build-dev": "set NODE_ENV=development && webpack --watch --colors --progress --devtool inline-source-map",
+		"build-prod": "set NODE_ENV=production && webpack --progress"
+	},
+	"repository": "https://github.com/Lord-Lavios/Torrent-Organizer.git",
+	"author": "Lord-Lavios",
+	"license": "MIT",
+	"description": "It organizes your movies and tv shows into their respective folders with posters. It also converts SDH subtitles to normal subtitles. It also supports hardlinking and symlinking for files.",
+	"dependencies": {
+		"babel-runtime": "^6.26.0"
+	},
+	"devDependencies": {
+		"babel-core": "^6.26.0",
+		"babel-eslint": "^8.0.3",
+		"babel-plugin-transform-runtime": "^6.23.0",
+		"babel-preset-env": "^1.6.1",
+		"uglifyjs-webpack-plugin": "^1.1.4",
+		"webpack": "^3.10.0"
+	}
 }
