@@ -1,22 +1,8 @@
 "use strict";
 
-const fs = require("fs");
-const HelperFuncs = require("./helper");
+import fs from "fs";
 
-const Helper = new HelperFuncs();
-
-module.exports = function () {
-
-	/*Get all the files in the provided path */
-	this.readFiles = basePath => {
-		let files = [];
-		fs.readdirSync(basePath).map(file => {
-			files.push(basePath + file);
-			if(!Helper.isDir(basePath + file)) return;
-			files = [...files, ...toTheDeepestFile({basePath, deeperDir: file, arr: []})];
-		});
-		return files;
-	};
+export default function (Helper) {
 
 	/* Gets all the files inside the folder */
 	function toTheDeepestFile({basePath, deeperDir, arr}) {
@@ -30,4 +16,21 @@ module.exports = function () {
 		return arr;
 	}
 
-};
+	class Reader {
+
+		/*Get all the files in the provided path */
+		readFiles(basePath) {
+			let files = [];
+			fs.readdirSync(basePath).map(file => {
+				files.push(basePath + file);
+				if(!Helper.isDir(basePath + file)) return;
+				files = [...files, ...toTheDeepestFile({basePath, deeperDir: file, arr: []})];
+			});
+			return files;
+		}
+		
+	}
+	
+	return new Reader();
+
+}
