@@ -1,60 +1,61 @@
 "use strict";
 
-const fs = require("fs");
-const HelperFuncs = require("./helper");
-const SubsFuncs = require("./fixSubs");
-const GetFilesFuncs = require("./getFiles.js");
+import fs from "fs";
 
-const Helper = new HelperFuncs();
-const Subs = new SubsFuncs();
-const GetFiles = new GetFilesFuncs();
+//Custom Functions
+import HelperFuncs from "./helper";
+import fixSubs from "./fixSubs";
+import GetFilesFuncs from "./getFiles";
+import parseArgs from "./parseArgs";
 
-/* Start of the Function */
+const Helper = HelperFuncs();
+const GetFiles = GetFilesFuncs(Helper);
+
 (async function () {
 	try {
-		if(!process.argv[2] || !process.argv[3]) { console.log("invalid path or mode"); return; }
-		let [basePath, link] = [process.argv[2].replace(/\\/g, "/"), process.argv[3]];
-		if(basePath[basePath.length - 1] !== "/") basePath += "/";
+		const args = await parseArgs();
+		if (!args) return;
+		let basePath = args.path;
 		console.time("It took");
 		console.log(`Organizing ${basePath}`);
-		console.log("Reading Files");
+		console.log("Reading Files...");
 		let files = GetFiles.readFiles(basePath);
 		console.log("Filtering Files into video, directories and other files");
 		let {dirs, video, other} = filterFiles(files);
 		console.log("Filtering movies and tv shows files");
-		let showsAndMovies = filterShowsAndMovies(video);
-		console.log("Getting shows and movies data from OmdbAPI.com");
-		let [showsData, posters, moviesData] = await apiShowsAndMovies(showsAndMovies);
-		let shows = Helper.replaceNameWithApiName({showsAndMovies, showsData});
+		let [shows, movies] = filterShowsAndMovies(video);
+		let [showsData, posters, moviesData] = [{}, {}, {}];
+		if (args.apiKey) {
+			console.log("Getting shows and movies data from OmdbAPI.com");
+			[showsData, posters, moviesData] = await apiShowsAndMovies([shows, movies]);
+			shows = Helper.replaceNameWithApiName({"showsAndMovies": [shows, movies], showsData});
+		}
 		console.log("Making new folders for movies and tv shows");
 		basePath += Helper.generateRandomFolderName();
-		await makeShowAndMoviesFolders({basePath, shows, posters, movies: moviesData});
+		await makeShowAndMoviesFolders({basePath, shows, posters, "movies": moviesData.length ? moviesData : movies});
 		console.log("Finding new names for movies and tv shows");
-		let newNames = findNewNamesForFiles({video, shows, showsData, moviesData});
-		if(link === "--symlink") {
-			console.log("Creating Symlinks");
-			newNames.map(({oldFile, newFile}) => fs.symlinkSync(oldFile, basePath + newFile));
-		} else if(link === "--hardlink") {
+		let newNames = findNewNamesForFiles({video, shows, showsData, movies, moviesData});
+		if(args.mode === 0) {
 			console.log("Creating Hardlinks");
-			newNames.map(({oldFile, newFile}) => fs.linkSync(oldFile, basePath + newFile));
-		} else if(link === "--no-link") {
+			newNames.map(({oldFile, newFile}) => fs.link(oldFile, basePath + newFile, err => err ? console.log(err): ""));
+		} else if(args.mode === 1) {
+			console.log("Creating Symlinks");
+			newNames.map(({oldFile, newFile}) => fs.symlink(oldFile, basePath + newFile, err => err ? console.log(err): ""));
+		} else if(args.mode === 2) {
 			console.log("Renaming Files");
-			newNames.map(({oldFile, newFile}) => fs.renameSync(oldFile, basePath + newFile));
+			newNames.map(({oldFile, newFile}) => fs.rename(oldFile, basePath + newFile, err => err ? console.log(err): ""));
 			await Promise.all(other.map(async file => await whatToDoWithFile(file, basePath))); //It will deal with all the srt, false positives in movies, and tv shows and other files
 			console.log("Deleting uneccesary files");
 			removeDirs(dirs);
-		} else {
-			console.log("Invalid mode");
 		}
 		console.log("Your organized files are in - " + basePath);
 		console.timeEnd("It took");
 	} catch(e) { console.log("Organize " + new Error(e)); }
 })();
 
-
-function findNewNamesForFiles({video, shows, showsData, moviesData}) {
+function findNewNamesForFiles({video, shows, showsData, movies, moviesData}) {
 	let names = [];
-	video.map(file => file.type === "movie" ? names.push(findNewNameForMovie(file, moviesData)) : "");
+	video.map(({file, name}) => file.type === "movie" ? names.push(findNewNameForMovie({file, name, moviesData, movies})) : "");
 	Object.keys(shows).map(name => names = [...names, ...findNewNameForShow({name, files: shows[name].files, showsData})]);
 	return names.filter(({newFile}) => newFile); //No API Match but pattern match
 }
@@ -64,7 +65,7 @@ function findNewNameForShow({name, files, showsData}) {
 	files.map(({file, episodeNum, season}) => {
 		let newFile = {oldFile: file};
 		let ext = Helper.getExt(file);
-		if(ext === ".srt") Subs.fixSubs(file);
+		if(ext === ".srt") fixSubs(file);
 		let title = Helper.getEpisodeTitle({name, episodeNum, season, showsData});
 		episodeNum = episodeNum < 10 ? "0" + episodeNum : episodeNum;
 		let baseName = `/Tv Shows/${name}/Season ${season}/${name} S${season < 10 ? "0" + season : season}E${episodeNum}`;
@@ -75,15 +76,20 @@ function findNewNameForShow({name, files, showsData}) {
 	return newFiles;
 }
 
-function findNewNameForMovie({file, name}, moviesData) {
+function findNewNameForMovie({file, name, moviesData, movies}) {
 	let newFile = {oldFile: file};
 	let ext = Helper.getExt(file);
-	if(ext === ".srt") Subs.fixSubs(file);
-	moviesData.map(item => {
-		if(name.toLowerCase() !== item.Title.toLowerCase()) return;
-		let {Title, Year, Runtime, Rating} = item;
-		newFile["newFile"] = `/Movies/${Title} ${Year} (${Runtime}) (${Rating})/${Title} ${Year}${ext}`;
-	});
+	if(ext === ".srt") fixSubs(file);
+	if(!moviesData.length) {
+		movies.forEach(movie => newFile["newFile"] = `/Movies/${movie}`);
+	} else {
+		moviesData.map(item => {
+			if(name.toLowerCase() !== item.Title.toLowerCase()) return;
+			let {Title, Year, Runtime, Rating} = item;
+			newFile["newFile"] = `/Movies/${Title} ${Year} (${Runtime}) (${Rating})/${Title} ${Year}${ext}`;
+		});
+	}
+	
 	return newFile;
 }
 
@@ -150,6 +156,7 @@ function filterShowsAndMovies(video) {
 		{
 			let {name, season, episodeNum} = fileStats;
 			if(!name) return;
+			name = Helper.capitalize(name);
 			let sameShow = Helper.sameShow(shows, name, season);
 			if(!sameShow) { shows[name] = {season: [season], length: 1, files: [{file, episodeNum, season}]}; return; } //New show detected
 			if(shows[sameShow.name] && shows[sameShow.name].hasOwnProperty("files")) shows[sameShow.name].files.push({file, episodeNum, season});
@@ -170,35 +177,33 @@ function removeDirs(files) {
 
 /* Makes folder for shows and movies */
 function makeShowAndMoviesFolders({basePath, shows, posters, movies}) {
-	try {
-		return new Promise(async resolve => {
-			fs.mkdirSync(basePath);
-			["Tv Shows", "Movies", "No Match Found"].map(str => fs.mkdirSync(`${basePath}/${str}`)); //Initial Folders
-			await Promise.all([makeShowsFolders({shows, basePath, posters}), makeMoviesFolders(movies, basePath)]);
-			resolve();
-		});
-	} catch(e) { console.log("Make Show Folders " + new Error(e)); }
+	return new Promise(async resolve => {
+		fs.mkdirSync(basePath);
+		["Tv Shows", "Movies", "No Match Found"].map(str => fs.mkdirSync(`${basePath}/${str}`)); //Initial Folders
+		await Promise.all([makeShowsFolders({shows, basePath, posters}), makeMoviesFolders(movies, basePath)]);
+		resolve();
+	}).catch(e => console.log("Make Show Folders " + new Error(e)));
 }
 
 /* Makes folder for the shows with; Season and showName */
 function makeShowsFolders({shows, posters, basePath}) {
-	try {
-		return new Promise(async resolve => {
-			for(let showName of Object.keys(shows)) {
-				let {season} = shows[showName];
-				fs.mkdirSync(`${basePath}/Tv Shows/${showName}`);
-				await savePosters({basePath, showName, posters});
-				season.map(season => fs.mkdirSync(`${basePath}/Tv Shows/${showName}/Season ${season}`));
-			}
-			resolve();
-		});
-	} catch(e) { console.log("makeShowsFolders "); console.log(new Error(e)); }
+	return new Promise(async resolve => {
+		for(let showName of Object.keys(shows)) {
+			let {season} = shows[showName];
+			fs.mkdirSync(`${basePath}/Tv Shows/${showName}`);
+			if (posters.length) await savePosters({basePath, showName, posters}); //API key is provided
+			season.map(season => fs.mkdirSync(`${basePath}/Tv Shows/${showName}/Season ${season}`));
+		}
+		resolve();
+	}).catch(e => console.log("makeShowsFolders " + new Error(e)));
 }
 
 /* Makes folder for the movies with name, year, rating and runtime */
 function makeMoviesFolders(movies, basePath) {
-	try {
-		return new Promise(async resolve => {
+	return new Promise(async resolve => {
+		if(movies instanceof Array) { //API key not provided
+			movies.forEach(movie => fs.mkdirSync(`${basePath}/Movies/${movie}`));
+		} else {
 			for(let movie of movies) {
 				let keys = Object.keys(movie);
 				keys.splice(2, 1); //Remove Poster
@@ -208,9 +213,9 @@ function makeMoviesFolders(movies, basePath) {
 				fs.mkdirSync(`${basePath}/Movies/${folder}`);
 				if(Poster !== "N/A") await Helper.saveImage(Poster, `${basePath}/Movies/${folder}/${Title}.jpg`);
 			}
-			resolve();
-		});
-	} catch(e) { console.log("makeMoviesFolders " + new Error(e)); }
+		}
+		resolve();
+	}).catch(e => console.log("makeMoviesFolders " + new Error(e)));
 }
 
 /* Downloads and save posters */
